fix: 修复 MinerU 跨平台 CI 检查
